requirements.txt
appdirs==1.4.4
argon2-cffi==20.1.0
asgiref==3.3.1
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bleach==3.3.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.7
django-ckeditor==6.0.0
django-extensions==3.1.1
django-heroku==0.3.1
django-js-asset==1.2.2
entrypoints==0.3
filelock==3.0.12
flake8==3.8.4
gunicorn==20.0.4
idna==2.10
ipykernel==5.5.0
ipython==7.16.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.11.3
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.11
jupyter-console==6.2.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
notebook==6.2.0
packaging==20.9
pandocfilters==1.4.3
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
pipenv==2020.11.15
prometheus-client==0.9.0
prompt-toolkit==3.0.16
ptyprocess==0.7.0
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
Pygments==2.8.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
pyzmq==22.0.3
qtconsole==5.0.2
QtPy==1.9.0
requests==2.25.1
Send2Trash==1.5.0
six==1.15.0
sqlparse==0.4.1
terminado==0.9.2
testpath==0.4.4
tornado==6.1
traitlets==5.0.5
urllib3==1.26.3
virtualenv==20.4.2
virtualenv-clone==0.5.4
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==5.2.0
widgetsnbextension==3.5.1

djangorestframework==3.12.2
drf-extensions==0.7.0
gunicorn==20.0.4